Fix compilation warnings under gcc13 and clang16
Summary:
Fix compilation warnings under gcc13 and clang16

Additionally fix FSR options in one test JSON file

Test Plan: All tests give identical results

Reviewers: jback, abudinen, kreps

Reviewed By: abudinen

Subscribers: #evtgen_admins

Tags: #evtgen

Differential Revision: https://phab.hepforge.org/D122

+++ EvtGenBase/EvtVectorParticle.hh
@@ -40,13 +40,13 @@ class EvtVectorParticle : public EvtParticle {
                const EvtVector4C&, const EvtVector4C& );
     EvtVector4C epsParent( int i ) const override
     {
-        assert( i >= 0 && i < m_eps.size() );
+        assert( i >= 0 && static_cast<std::size_t>( i ) < m_eps.size() );
 
         return boostTo( m_eps[i], this->getP4() );
     }
     EvtVector4C eps( int i ) const override
     {
-        assert( i >= 0 && i < m_eps.size() );
+        assert( i >= 0 && static_cast<std::size_t>( i ) < m_eps.size() );
 
         return m_eps[i];
     }
